backend/requirements.txt
fastapi 
uvicorn[standard] 
pymongo[srv]
python-dotenv
google-generativeai
requests
pydantic
vosk
pyaudio
pygame
openai
